requirements.txt
# Core dependencies for Lie-Circuit experiment
transformer-lens==1.11.0
torch==2.2.1
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
hydra-core>=1.3.0
openai>=1.0.0
scikit-learn>=1.3.0  # For Cohen's kappa
networkx>=3.0  # For circuit visualization
seaborn>=0.12.0  # For plotting
jupyter>=1.0.0
ipywidgets>=8.0.0
datasets>=2.14.0  # For HuggingFace datasets
python-dotenv>=1.0.0  # For environment variables
scipy>=1.10.0  # For statistical analysis